:test_tube: test waitingRoom
