+++ frontend/src/pages/games/WaitingRoom.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from "react";
 import { useHistory, useParams } from "react-router-dom"
 import { socket } from "../../socket/socket";
 import { User } from "../../types/userTypes";
-import { gameRoomDetail } from "../../types/gameTypes"
-import { gameRoomInitialState, updateGameRoom } from "../../redux/gameReducer";
+import { gameRoomDetail, GameUser } from "../../types/gameTypes"
+import { updateGameRoom, updatePlayRoom } from "../../redux/gameReducer";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../redux/rootReducer";
 import "./WaitingRoom.css"
@@ -22,11 +22,50 @@ export default function WaitingRoom(){
 		console.log("waitingRoom");
 		if (param.id && param.id !== room.roomid){
 			socket.emit("exitGameRoom", {roomid: room.roomid});
-			dispatch(updateGameRoom(gameRoomInitialState));
-			setRoom(gameRoomInitialState);
-			history.push("/game");
-			// socket.emit("gameRoomList");
-			window.location.reload();
+		}
+		socket.on("changeGameRoom", (msg:any) => {
+			const tmp:gameRoomDetail = room;
+			console.log("changeGameRoom");
+			console.log(msg);
+			if (msg.manager) {tmp.manager = msg.manager;}
+			if (msg.title) {tmp.title = msg.title;}
+			if (msg.speed) {tmp.speed = msg.speed;}
+			if (msg.status) {tmp.status = msg.status;}
+			if (msg.type) {tmp.type = msg.type;}
+			if (msg.addObserver) {
+				const observer:GameUser = msg.addObserver;
+				const idx:number = tmp.observer.findIndex((person:GameUser)=>person.userid===observer.userid);
+				if (idx === -1){ tmp.observer.push(observer) }
+			}
+			if (msg.deleteObserver) {
+				const observer:GameUser = msg.deleteObserver;
+				tmp.observer = tmp.observer?.filter((ob: GameUser) => ob.userid !== observer.userid);
+			}
+			if (msg.addPlayer) {
+				const player:GameUser = msg.addPlayer;
+				const idx:number = tmp.players.findIndex((person:GameUser)=>person.userid === player.userid);
+				if (idx === -1){ tmp.players.push(player); }
+			}
+			if (msg.deletePlayer) {
+				const player:GameUser = msg.deletePlayer;
+				tmp.players = tmp.players?.filter((person: GameUser) => person.userid !== player.userid);
+			}
+			setRoom({...tmp});
+			dispatch(updateGameRoom(tmp));
+		});
+		socket.on("startGame", (msg:any) => {
+			console.log("start game!");
+			console.log(msg);
+			if (msg.result) {
+				alert("failed to play the game!");
+			} else {
+				dispatch(updatePlayRoom(msg));
+				history.push(`/game/play/${msg.roomid}`);
+			}
+		});
+		return ()=>{
+			socket.off("changeGameRoom");
+			socket.off("startGame");
 		}
 	}, [dispatch, gameroom, room, history, param]);
